refactor: move verification schema to core
