+++ packages/better-auth/src/types/models.ts
@@ -1,10 +1,8 @@
 import type { BetterAuthOptions } from "./options";
-import type { verificationSchema } from "../db/schema";
 import type { Auth } from "../auth";
 import type { InferFieldsFromOptions, InferFieldsFromPlugins } from "../db";
 import type { StripEmptyObjects, UnionToIntersection } from "./helper";
 import type { BetterAuthPlugin } from "./plugins";
-import type * as z from "zod";
 import type { User, Session } from "@better-auth/core/db";
 
 export type Models =
@@ -85,6 +83,10 @@ interface RateLimit {
 	lastRequest: number;
 }
 
-export type { User, Account, Session } from "@better-auth/core/db";
-export type Verification = z.infer<typeof verificationSchema>;
+export type {
+	User,
+	Account,
+	Session,
+	Verification,
+} from "@better-auth/core/db";
 export type { RateLimit };
