+++ packages/core/src/db/index.ts
@@ -10,6 +10,7 @@ export { coreSchema } from "./schema/shared";
 export { userSchema, type User } from "./schema/user";
 export { accountSchema, type Account } from "./schema/account";
 export { sessionSchema, type Session } from "./schema/session";
+export { verificationSchema, type Verification } from "./schema/verification";
 
 export type {
 	DBFieldAttribute,
